Fix "expired link" error
When an email address is verified a message is displayed and time-out set to take the user back to the previous route.

If the user changes route before this timeout then, when the timeout fires, it will try to redirect back to /verify-email. This causes a 2nd token verification and thus generates an "Expired" message.

We should never redirect to the /verify-email route. This fixes that.