atInput: gather Field object from current data
Previously, atInput template assumed 'this' in event handlers
was set to the Field instance the parent template set as our
data context.

This is true for some of the simplest input types like 'text',
'checkbox' and other. However, in the general case this is
not the case.

From Meteor's documentation:

    Event Maps
    ...
    The handler also receives some additional context data
    in this, depending on the context of the current element
    handling the event.

Thus, 'this' in event handlers refer to the context of the
element the event was fired on.

And the 'radio' input type is problematic because the input
element is inside the context provided by a 'each' template.

Meteor's documentation also says:

    Template.currentData()
    ...
    Inside an event handler, returns the data context of the
    template on which this event handler was defined.

And given that handlers are defined at the atInput template what
we really want is Template.currentData().

For instance materialize's at_input.js reads

    Template.atInput.helpers(AccountsTemplates.atInputHelpers);

and so do other styles.

So, as long as that is the case, the field object instance will
be available at Template.currentData(), which is what this patch
proposes to use.

Fixes 473